bug fix on NPC size
Previously we assumed the NPC's only took up one tile, however a lot of attackable NPC's take up more than 1 tile, for example green dragons, moss giants, most bosses, etc.. We now account for the size of the NPC and calculate the most dense tiles based on all the tiles that the NPC's are taking up. This code does assume that the NPC's are all square, which seems to be the case is nearly every instance however there may be a few cases where this is not true.